64-bit, 10.5
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2804 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DragMatrix.m
@@ -331,7 +331,7 @@ NSString *pasteBoardTypeCover = @"KeyImages";
 *****************************************************************************/
 - (void)mouseUp:(NSEvent *)event {
     BOOL cellWasHit;
-    int row, column;
+    NSInteger row, column;
     NSPoint point;
     
     point = [self convertPoint:[event locationInWindow] fromView:nil];
@@ -368,7 +368,7 @@ NSString *pasteBoardTypeCover = @"KeyImages";
  * If we hit a cell, then start the drag 
 *****************************************************************************/
 - (void)mouseDragged:(NSEvent *)event { 
-    int row, column; 
+    NSInteger row, column; 
     NSPoint point;
 
     point = [self convertPoint:[event locationInWindow] fromView:nil];
